+++ src/Limelight-internal.h
@@ -43,7 +43,7 @@ extern SS_PING VideoPingPayload;
 
 #define SS_FF_PEN_TOUCH_EVENTS 0x01
 #define SS_FF_CONTROLLER_TOUCH_EVENTS 0x02
-extern uint16_t SunshineFeatureFlags;
+extern uint32_t SunshineFeatureFlags;
 
 #ifndef UINT24_MAX
 #define UINT24_MAX 0xFFFFFF
